chore: drive-by simplify test files
