+++ rust/tests/command_vacuum.rs
@@ -1,13 +1,40 @@
 use chrono::Duration;
 use common::clock::TestClock;
-use common::schemas::{get_vacuum_underscore_schema, get_xy_date_schema};
 use common::TestContext;
 use deltalake::operations::vacuum::Clock;
 use deltalake::operations::DeltaOps;
+use deltalake::Schema;
 use object_store::{path::Path, Error as ObjectStoreError, ObjectStore};
+use serde_json::json;
 use std::sync::Arc;
 
 mod common;
+
+/// Basic schema
+pub fn get_xy_date_schema() -> Schema {
+    serde_json::from_value(json!({
+      "type": "struct",
+      "fields": [
+        {"name": "x", "type": "integer", "nullable": false, "metadata": {}},
+        {"name": "y", "type": "integer", "nullable": false, "metadata": {}},
+        {"name": "date", "type": "string", "nullable": false, "metadata": {}},
+      ]
+    }))
+    .unwrap()
+}
+
+/// Schema that contains a column prefiexed with _
+pub fn get_vacuum_underscore_schema() -> Schema {
+    serde_json::from_value::<Schema>(json!({
+      "type": "struct",
+      "fields": [
+        {"name": "x", "type": "integer", "nullable": false, "metadata": {}},
+        {"name": "y", "type": "integer", "nullable": false, "metadata": {}},
+        {"name": "_date", "type": "string", "nullable": false, "metadata": {}},
+      ]
+    }))
+    .unwrap()
+}
 
 #[tokio::test]
 // Validate vacuum works on a non-partitioned table
